Memory/paging fixes, new elf loader
* Laying the groundwork for a scheduler this commit fixes a whole bunch
  of issues related to paging and VMM.
* Simplify and improve load_elf(...) and parse/elf.
* Add the ability to debug memory management
* Some smaller style and syntax fixes

+++ libs/alloc/bitmap.h
@@ -7,7 +7,7 @@
 #define BITMAP_FREE 0
 #define BITMAP_USED 1
 
-#define ALLOC_OUT_OF_BLOCKS ((usize)(-1))
+#define ALLOC_OUT_OF_BLOCKS (-1)
 
 typedef struct {
     void* chuck_start;
